Issue #118. HOC applied. 	new file:   src/components/HOCs/withConfirm.tsx 	new file:   src/components/HOCs/withSnackbar.tsx 	modified:   src/components/LevelView/LevelCat.tsx 	modified:   src/components/LevelView/LevelCatEdit.tsx 	modified:   src/components/LevelView/index.tsx 	modified:   src/components/ProjectDetailView/ProjectLevels.tsx 	modified:   src/components/ProjectDetailView/ProjectLevelsWrapper.tsx 	deleted:    src/components/shared/ConfirmDialog.js 	new file:   src/components/shared/ConfirmDialog.tsx 	deleted:    src/containers/GenContractorView/AddProjectView/Levels.tsx 	modified:   src/containers/GenContractorView/AddProjectView/index.tsx 	modified:   src/containers/GenContractorView/CurrentProjectView/index.tsx 	modified:   src/types/project.d.ts

+++ src/components/ProjectDetailView/ProjectLevels.tsx
@@ -18,6 +18,7 @@ import { Theme, Typography } from '@material-ui/core';
 
 import TabPanel from 'components/TabPanel';
 import LevelView from 'components/LevelView';
+import withConfirm, { withConfirmProps } from 'components/HOCs/withConfirm';
 
 import { ProjectLevel, ProjectLevelCategory } from 'types/project';
 
@@ -60,7 +61,7 @@ export interface IProjectLevelsProps extends RouteComponentProps {
     deleteCategory: (lvlId: number, catId: number) => void;
 }
 
-const ProjectLevels: React.SFC<IProjectLevelsProps> = props => {
+const ProjectLevels: React.SFC<IProjectLevelsProps & withConfirmProps> = props => {
 
     const {
         levels,
@@ -69,6 +70,8 @@ const ProjectLevels: React.SFC<IProjectLevelsProps> = props => {
         updateCategory,
         addCategory,
         deleteCategory,
+        showConfirm,
+        hideConfirm
     } = props;
     const classes = useStyles({});
 
@@ -91,8 +94,11 @@ const ProjectLevels: React.SFC<IProjectLevelsProps> = props => {
     const showDialog = () => setModal(true);
 
     const deleteLvl = (id) => {
-        deleteLevel(id);
-        if (value > 0) setValue(value - 1);
+        showConfirm('Confirm', 'Do you really want to delete this level?', () => {
+            hideConfirm();
+            deleteLevel(id);
+            if (value > 0) setValue(value - 1);
+        }, true);
     }
 
     return (
@@ -193,4 +199,4 @@ const ProjectLevels: React.SFC<IProjectLevelsProps> = props => {
     );
 }
 
-export default ProjectLevels;
+export default withConfirm<IProjectLevelsProps>(ProjectLevels);
